Rename interface->configuration, connectivity->interface
