+++ src/device/rdk.rs
@@ -2,10 +2,10 @@ pub mod applications;
 pub mod device;
 pub mod health_check;
 pub mod input;
-pub mod interface;
+pub mod configuration;
 pub mod operations;
 pub mod output;
 pub mod system;
 pub mod version;
 pub mod voice;
-pub mod connectivity;
+pub mod interface;
