--- src/device/rdk/connectivity.rs
@@ -1,3 +1,0 @@
-pub mod ws;
-pub mod http;
-pub mod rdk;
