+++ src/device/rdk/device/info.rs
@@ -6,10 +6,10 @@ use crate::dab::structs::DisplayType;
 use crate::dab::structs::GetDeviceInformationResponse;
 use crate::dab::structs::NetworkInterface;
 use crate::dab::structs::NetworkInterfaceType;
-use crate::device::rdk::interface::get_device_id;
-use crate::device::rdk::connectivity::http::http_post;
-use crate::device::rdk::connectivity::rdk::get_thunder_property;
-use crate::device::rdk::interface::get_rdk_device_info;
+use crate::device::rdk::configuration::get_device_id;
+use crate::device::rdk::interface::http::http_post;
+use crate::device::rdk::interface::rdk::get_thunder_property;
+use crate::device::rdk::configuration::get_rdk_device_info;
 use serde::{Deserialize, Serialize};
 
 #[allow(non_snake_case)]
