+++ src/device/rdk/system/settings/list.rs
@@ -7,14 +7,14 @@ use crate::dab::structs::ListSystemSettingsResponse;
 use crate::dab::structs::MatchContentFrameRate;
 use crate::dab::structs::OutputResolution;
 use crate::dab::structs::VideoInputSource;
-use crate::device::rdk::connectivity::rdk::rdk_request;
-use crate::device::rdk::connectivity::rdk::rdk_request_with_params;
-use crate::device::rdk::connectivity::rdk::rdk_sound_mode_to_dab;
-use crate::device::rdk::connectivity::rdk::service_is_available;
-use crate::device::rdk::connectivity::rdk::RdkResponse;
+use crate::device::rdk::interface::rdk::rdk_request;
+use crate::device::rdk::interface::rdk::rdk_request_with_params;
+use crate::device::rdk::interface::rdk::rdk_sound_mode_to_dab;
+use crate::device::rdk::interface::rdk::service_is_available;
+use crate::device::rdk::interface::rdk::RdkResponse;
 use crate::device::rdk::system::settings::get::get_rdk_audio_port;
-use crate::hw_specific::interface::get_audio_volume_range;
-use crate::hw_specific::interface::get_supported_languages;
+use crate::hw_specific::configuration::get_audio_volume_range;
+use crate::hw_specific::configuration::get_supported_languages;
 use lazy_static::lazy_static;
 use serde::{Deserialize, Serialize};
